fix: Update text and layout for step 0
