+++ src/components/common/ConnectWallet/WalletDetails.tsx
@@ -1,28 +1,12 @@
 import { Button, Typography } from '@mui/material'
 import type { ReactElement } from 'react'
 
-import useOnboard, { connectWallet } from '@/hooks/wallets/useOnboard'
-import { OVERVIEW_EVENTS } from '@/services/analytics/events/overview'
 import KeyholeIcon from '@/components/common/icons/KeyholeIcon'
-import { trackEvent } from '@/services/analytics'
-import { CodedException } from '@/services/exceptions'
 import type { ConnectedWallet } from '@/services/onboard'
+import useConnectWallet from '@/components/common/ConnectWallet/useConnectWallet'
 
 const WalletDetails = ({ onConnect }: { onConnect?: (wallet?: ConnectedWallet) => void }): ReactElement => {
-  const onboard = useOnboard()
-
-  const handleConnect = async () => {
-    if (!onboard) return
-
-    // We `trackEvent` instead of using `<Track>` as it impedes styling
-    trackEvent(OVERVIEW_EVENTS.OPEN_ONBOARD)
-
-    const result = await connectWallet(onboard)
-
-    if (result instanceof CodedException) return
-
-    onConnect?.(result)
-  }
+  const handleConnect = useConnectWallet(onConnect)
 
   return (
     <>
